fix: refactory
delete darkmode in GlobalContext
Use Dark classes to change images in Navbar and Footer
