+++ src/components/Logged/Events/CreateEvents.jsx
@@ -1,5 +1,4 @@
 import { useContext, useEffect, useState } from "react";
-import { GlobalContext } from "../../../context/GlobalContext";
 import {
   Popover,
   EventCard,
