+++ src/context/GlobalContext.jsx
@@ -13,10 +13,6 @@ import { getAuth } from "firebase/auth";
 export const GlobalContext = createContext();
 
 export function GlobalContextProvider(props) {
-  const [darkMode, setDarkMode] = useState(true);
-  const [navbar, setNavbar] = useState(false);
-  const [footer, setFooter] = useState(false);
-
   const firebaseConfig = {
     apiKey: "AIzaSyAvYBiGiqVlunBsuOhheOhsQ5iR3l60cks",
     authDomain: "blocksidian-7b83e.firebaseapp.com",
@@ -32,22 +28,6 @@ export function GlobalContextProvider(props) {
   const analytics = getAnalytics(app);
   const auth = getAuth();
   const db = getFirestore(app);
-
-  const useDarkMode = () => {
-    const [enabled, setEnabled] = useLocalStorage("dark-theme");
-    const isEnabled = typeof enabled === "undefined" ? enabled : enabled;
-
-    useEffect(() => {
-      const className = "dark";
-      const bodyClass = window.document.body.classList;
-
-      isEnabled ? bodyClass.add(className) : bodyClass.remove(className);
-
-      setDarkMode(enabled);
-    }, [enabled, isEnabled]);
-
-    return [enabled, setEnabled];
-  };
 
   const [username, setUsername] = useState("Username");
   const [userId, setUserId] = useState("");
@@ -241,12 +221,6 @@ export function GlobalContextProvider(props) {
   return (
     <GlobalContext.Provider
       value={{
-        useDarkMode,
-        darkMode,
-        setNavbar,
-        navbar,
-        setFooter,
-        footer,
         globalEvents,
         myEvents,
         popularEvents,
@@ -261,34 +235,4 @@ export function GlobalContextProvider(props) {
       {props.children}
     </GlobalContext.Provider>
   );
-}
-
-function useLocalStorage(key, initialValue) {
-  // Guardamos el valor de una llave en el localstorage
-  // Si existe hace todo esto
-  const [storedValue, setStoredValue] = useState(() => {
-    try {
-      const item = window.localStorage.getItem(key);
-      return item ? JSON.parse(item) : initialValue;
-    } catch (error) {
-      console.log(error);
-      return initialValue;
-    }
-  });
-
-  // Si no existe la agrega al localstorage
-
-  const setValue = (value) => {
-    try {
-      const ValueToStore =
-        value instanceof Function ? value(storedValue) : value;
-
-      setStoredValue(ValueToStore);
-
-      window.localStorage.setItem(key, JSON.stringify(ValueToStore));
-    } catch (error) {
-      console.log(error);
-    }
-  };
-  return [storedValue, setValue];
 }
